Separate movement functions in a new file movements.c
