+++ sources/drawing_part.c
@@ -6,17 +6,19 @@
 /*   By: dbouron <dbouron@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/01 18:43:02 by dbouron           #+#    #+#             */
-/*   Updated: 2022/09/06 18:20:36 by dbouron          ###   ########lyon.fr   */
+/*   Updated: 2022/09/07 17:41:16 by dbouron          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
 
-void	draw_in_image(t_mlx_params *mlx, t_image *image, t_raycasting *raycasting)
+void	draw_in_image(t_mlx *mlx, t_image *image, t_raycasting *raycasting)
 {
 	mlx_clear_window(mlx->mlx, mlx->window);
 	create_image(mlx, image);
 	raycasting_algo(image, raycasting);
+	mlx_put_image_to_window(mlx->mlx, mlx->window, image->img, 0, 0);
+	mlx_destroy_image(mlx->mlx, image->img);
 }
 
 void	my_img_pixel_put(t_image *image, int x, int y, int color)
