+++ sources/initializing.c
@@ -6,7 +6,7 @@
 /*   By: dbouron <dbouron@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/25 10:58:07 by dbouron           #+#    #+#             */
-/*   Updated: 2022/09/07 15:29:21 by dbouron          ###   ########lyon.fr   */
+/*   Updated: 2022/09/07 17:39:20 by dbouron          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,6 @@ void	init_raycasting_values(t_raycasting *raycasting)
 	raycasting->direction_y = 0;
 	raycasting->camera_plane_x = 0;
 	raycasting->camera_plane_y = 0.66;
-	raycasting->move_speed = 0.5;
+	raycasting->move_speed = 0.2;
 	raycasting->rot_speed = 0.1;
-
 }
