+++ sources/main.c
@@ -6,7 +6,7 @@
 /*   By: dbouron <dbouron@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/03 15:33:54 by dbouron           #+#    #+#             */
-/*   Updated: 2022/09/06 17:06:00 by dbouron          ###   ########lyon.fr   */
+/*   Updated: 2022/09/07 17:43:30 by dbouron          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 int	main(int argc, char **argv)
 {
 	(void) argv; //à enlever
-	if (argc == 1)//changer a 2 quand on aura le parsing
+	if (argc == 1) //changer a 2 quand on aura le parsing
 	{
 		display_window();
 	}
